마트 정보 저장/삭제 스케줄러 등록
