+++ src/main/java/com/project/recipick/controller/DataController.java
@@ -11,7 +11,6 @@ import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
@@ -165,90 +164,5 @@ public class DataController {
         String addr = "http://211.237.50.150:7080/openapi/";
         recipeCardService.callRecipeIngredientAndSaveData(apiKey, addr);
     }
-
-    @GetMapping("/saveMartInfo")
-    public void saveMartInfo() throws ParseException {
-        String apiKey =  martInfoApiKey;
-
-        int pageSize = 1000;
-        int start = 1;
-        int lastEnd = 0;
-
-        while (true) {
-            int end = start + pageSize - 1;
-            lastEnd = end;
-
-            String url = "http://openAPI.seoul.go.kr:8088/" + apiKey + "/json/ListNecessariesPricesService/" + start + "/" + end + "/?P_YEAR_MONTH=2025-06";
-
-            ResponseEntity<String> responseEntity = restTemplate.getForEntity(url, String.class);
-            String responseBody = responseEntity.getBody();
-
-            String jsonString = responseBody;
-            JSONParser parser = new JSONParser();
-
-            JSONObject jsonObject = (JSONObject) parser.parse(jsonString);
-            String ListNecessariesPricesService = jsonObject.get("ListNecessariesPricesService").toString();
-
-            if (ListNecessariesPricesService == null) break; // 서비스 자체가 없으면 종료
-
-
-            JSONObject jsonObj = (JSONObject) parser.parse(ListNecessariesPricesService);
-            String row = jsonObj.get("row").toString();
-
-            if (row == null) break; // 서비스 자체가 없으면 종료
-
-            JSONArray jsonArray = (JSONArray) parser.parse(row);
-
-            if (jsonArray.isEmpty()) break; // 빈 배열이면 종료
-
-            boolean hasTargetMonth = false;
-
-            System.out.println(jsonArray.get(0));
-
-            for(int i=0; i<jsonArray.size(); i++) {
-                JSONObject obj = (JSONObject) jsonArray.get(i);
-
-                // 가격이 '0'이면 저장하지 않음
-                String price = obj.get("A_PRICE").toString();
-                if ("0".equals(price)) continue;
-
-                String pYearMonth = obj.get("P_YEAR_MONTH").toString();
-                if ("2025-06".equals(pYearMonth)) {
-                    hasTargetMonth = true;
-
-                    MartInfo martInfo = new MartInfo();
-
-                    martInfo.setpSeq(obj.get("P_SEQ").toString());
-                    martInfo.setmSeq(obj.get("M_SEQ").toString());
-                    martInfo.setmName(obj.get("M_NAME").toString());
-                    martInfo.setaSeq(obj.get("A_SEQ").toString());
-                    martInfo.setaName(obj.get("A_NAME").toString());
-                    martInfo.setaUnit(obj.get("A_UNIT").toString());
-                    martInfo.setaPrice(price);
-                    martInfo.setpYearMonth(obj.get("P_YEAR_MONTH").toString());
-                    martInfo.setAddCol(obj.get("ADD_COL").toString());
-                    martInfo.setpDate(obj.get("P_DATE").toString());
-                    martInfo.setmTypeCode(obj.get("M_TYPE_CODE").toString());
-                    martInfo.setmTypeName(obj.get("M_TYPE_NAME").toString());
-                    martInfo.setmGuCode(obj.get("M_GU_CODE").toString());
-                    martInfo.setmGuName(obj.get("M_GU_NAME").toString());
-
-                    System.out.println("start & end = " + start + ", " + lastEnd);
-                    martInfoService.saveMartInfo(martInfo);
-                }
-
-            }
-
-            // 이번 페이지에 2025-06이 없으면 반복 종료
-            if (!hasTargetMonth) break;
-
-            // 다음 페이지로 이동
-            start += pageSize;
-        }
-
-
-        System.out.println("endpage = " + lastEnd);
-    }
-
 
 }
